+++ sync-for-payables/RELEASES.md
@@ -59,3 +59,13 @@ Based on:
 - [python v2.1.0] sync-for-payables
 ### Releases
 - [PyPI v2.1.0] https://pypi.org/project/codat-sync-for-payables/2.1.0 - sync-for-payables
+
+## 2023-10-30 01:24:22
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Payables.yaml
+- Speakeasy CLI 1.109.0 (2.173.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v2.1.1] sync-for-payables
+### Releases
+- [PyPI v2.1.1] https://pypi.org/project/codat-sync-for-payables/2.1.1 - sync-for-payables